move routes outside pages
